move files to services folder and update dependencies
