+++ src/miner.cpp
@@ -34,7 +34,7 @@
 #include <boost/tuple/tuple.hpp>
 #include <queue>
 #include <utility>
-#include "graph.h"
+#include "services/graph.h"
 using namespace std;
 //////////////////////////////////////////////////////////////////////////////
 //
