+++ src/services/asset.h
@@ -10,7 +10,7 @@
 #include "script/script.h"
 #include "serialize.h"
 #include "primitives/transaction.h"
-#include "assetallocation.h"
+#include "services/assetallocation.h"
 class CWalletTx;
 class CTransaction;
 class CReserveKey;
