+++ src/services/cert.h
@@ -9,7 +9,7 @@
 #include "dbwrapper.h"
 #include "script/script.h"
 #include "serialize.h"
-#include "assetallocation.h"
+#include "services/assetallocation.h"
 class CWalletTx;
 class CTransaction;
 class CReserveKey;
