+++ src/test/syscoin_asset_allocation_tests.cpp
@@ -6,16 +6,16 @@
 #include "utiltime.h"
 #include "util.h"
 #include "rpc/server.h"
-#include "alias.h"
-#include "cert.h"
-#include "asset.h"
+#include "services/alias.h"
+#include "services/cert.h"
+#include "services/asset.h"
 #include "base58.h"
 #include "chainparams.h"
 #include <boost/test/unit_test.hpp>
 #include <boost/lexical_cast.hpp>
 #include <iterator>
 #include <chrono>
-#include "ranges.h"
+#include "services/ranges.h"
 using namespace boost::chrono;
 using namespace std;
 BOOST_GLOBAL_FIXTURE( SyscoinTestingSetup );
