+++ src/test/test_syscoin_services.cpp
@@ -8,12 +8,12 @@
 #include "amount.h"
 #include "rpc/server.h"
 #include "feedback.h"
-#include "cert.h"
-#include "asset.h"
-#include "assetallocation.h"
-#include "escrow.h"
-#include "offer.h"
-#include "alias.h"
+#include "services/cert.h"
+#include "services/asset.h"
+#include "services/assetallocation.h"
+#include "services/escrow.h"
+#include "services/offer.h"
+#include "services/alias.h"
 #include "wallet/crypter.h"
 #include "random.h"
 #include "base58.h"
@@ -28,7 +28,7 @@
 #include <boost/thread.hpp>
 #include <boost/lexical_cast.hpp>
 #include <boost/algorithm/string/case_conv.hpp> // for to_upper()
-#include "ranges.h"
+#include "services/ranges.h"
 static int node1LastBlock = 0;
 static int node2LastBlock = 0;
 static int node3LastBlock = 0;
