+++ src/validation.cpp
@@ -55,13 +55,13 @@
 #include <boost/thread.hpp>
 // SYSCOIN
 #include "auxpow.h"
-#include "offer.h"
-#include "cert.h"
-#include "alias.h"
-#include "asset.h"
-#include "assetallocation.h"
-#include "escrow.h"
-#include "graph.h"
+#include "services/offer.h"
+#include "services/cert.h"
+#include "services/alias.h"
+#include "services/asset.h"
+#include "services/assetallocation.h"
+#include "services/escrow.h"
+#include "services/graph.h"
 #include "base58.h"
 #include "rpc/server.h"
 #include "thread_pool/thread_pool.hpp"
